+++ test/event-emitter.js
@@ -35,6 +35,15 @@ describe('EventEmitter', function () {
     expect(spy.callCount).to.equal(0);
   });
 
+  it('does not error when unsubscribing from an event that does not exist', function () {
+    var emitter = new EventEmitter();
+    var spy = sinon.stub();
+
+    expect(function () {
+      emitter.off('foo', spy);
+    }).to.not.throw();
+  });
+
   it('only unsubscribes from specific function', function () {
     var emitter = new EventEmitter();
     var spy1 = sinon.stub();
@@ -85,6 +94,19 @@ describe('EventEmitter', function () {
     }).to.throw('danger zone!');
 
     expect(thirdCallback).not.to.have.beenCalled;
+  });
+
+  it('can check if event has listeners', function () {
+    var emitter = new EventEmitter();
+    var noop = sinon.stub();
+
+    emitter.on('a', noop);
+    emitter.on('b', noop);
+    emitter.off('b', noop);
+
+    expect(emitter.hasListener('a')).to.equal(true);
+    expect(emitter.hasListener('b')).to.equal(false);
+    expect(emitter.hasListener('c')).to.equal(false);
   });
 
   describe('createChild', function () {