+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# unreleased
+
+* Fix issue where `off` would throw an error if no events existed for named event
+* Add `hasListener` method
+
 # 0.2.0
 
 * Add `off` method